Added an additional graph for december
Adding an additional graph for december to It occasionally approaches 100 and briefly drops to zero. Our data appears to go above 100% in this particular instance. Understanding how the sensors function requires visiting and speaking with the experiment design team. Potentially, the sensor is reading data that has been piled above a maximum fill line, which is set at 100%.